hyperon_common

Module multitrie

Multi-value trie with double side matching.

Key of the trie is a sequence of tokens. The most common token type is an TrieToken::Exact token. It contains a value and matches with another TrieToken::Exact token having the equal value. For example key [ Exact(A), Exact(B) ] matches with the equal key but doesn’t match with [ Exact(A), Exact(C) ]. But TrieKey containing TrieToken::Exact token is not the only kind of keys supported. MultiTrie also supports wildcards and sub-expressions.

Subexpression is a part of the key between a pair of left and right parentheses. TrieToken::LeftPar and TrieToken::RightPar are also kinds of tokens. Parentheses mark the edges of the key subsequence which can be matched as a whole. Exact token doesn’t match with a part of a subexpression unless it is also a part of a subexpression and located on the same position as a matched token. Subexpressions can be nested and in such case parentheses should be correctly balanced. For example [ Exact(A), LeftPar, Exact(B), RightPar ] is a key which contains a subexpression. It can be matched with the same key but doesn’t match with [ Exact(A), Exact(B) ].

The TrieToken::Wildcard is a last kind of token which matches the exact value, the whole sub-expression or another wildcard. It can recognize anything but when dealing with subexpression it matches with the whole subexpression only. It cannot be matched with a part of the subexpression or with the left or right parenthesis. For example [ Exact(A), * ] key matches both [ Exact(A), Exact(B) ] and [ Exact(A), LeftPar, Exact(B), RightPar ] keys.

Wildcard can be used not only for getting value but also as a key for keeping value. Thus a single key in the MultiTrie can match many different keys for retrieve value. In the example above we could put two values with [ Exact(A), Exact(B) ] and [ Exact(A), LeftPar, Exact(B), RightPar ] keys into the trie and then get both of them using [ Exact(A), * ] key. Or vice versa put a single value using [ Exact(A), * ] key and extract it using two keys above. This is what is called double-side matching.

Because of double-side matching MultiTrie cannot guarantie that single key corresponds to the single value. Thus MultiTrie::get call returns the iterator through the values which keys are matched with given key.

Structs

MultiTrie

Multi-value trie with double side matching. See crate::multitrie for the algorithm description.

TrieKey

Trie key is a sequence of TrieToken.

Enums

TrieToken

Single token of TrieKey. Each kind of token has its own recognition rules.